linkedin_recruting/cv_batch_processor/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
h11==0.14.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
onnxruntime==1.19.2
onnxruntime-genai==0.4.0
packaging==24.2
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
pyreadline3==3.5.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0